for bitwise use np.integer

+++ tests/test_series.py
@@ -1316,29 +1316,30 @@ def test_bitwise_operators() -> None:
     s = pd.Series([1, 2, 3, 4], dtype=int)
     s2 = pd.Series([9, 10, 11, 12], dtype=int)
     # for issue #348 (bitwise operators on Series should support int)
-    check(assert_type(s & 3, "pd.Series[int]"), pd.Series, np.intc)
-    check(assert_type(3 & s, "pd.Series[int]"), pd.Series, np.intc)
+    # The bitwise integers return platform-dependent numpy integers in the Series
+    check(assert_type(s & 3, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(3 & s, "pd.Series[int]"), pd.Series, np.integer)
 
-    check(assert_type(s | 3, "pd.Series[int]"), pd.Series, np.intc)
-    check(assert_type(3 | s, "pd.Series[int]"), pd.Series, np.intc)
+    check(assert_type(s | 3, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(3 | s, "pd.Series[int]"), pd.Series, np.integer)
 
-    check(assert_type(s ^ 3, "pd.Series[int]"), pd.Series, np.intc)
-    check(assert_type(3 ^ s, "pd.Series[int]"), pd.Series, np.intc)
+    check(assert_type(s ^ 3, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(3 ^ s, "pd.Series[int]"), pd.Series, np.integer)
 
     check(assert_type(s & [1, 2, 3, 4], "pd.Series[bool]"), pd.Series, np.bool_)
     check(assert_type([1, 2, 3, 4] & s, "pd.Series[bool]"), pd.Series, np.bool_)
-    check(assert_type(s & s2, "pd.Series[int]"), pd.Series, np.int_)
-    check(assert_type(s2 & s, "pd.Series[int]"), pd.Series, np.int_)
+    check(assert_type(s & s2, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(s2 & s, "pd.Series[int]"), pd.Series, np.integer)
 
     check(assert_type(s | [1, 2, 3, 4], "pd.Series[bool]"), pd.Series, np.bool_)
     check(assert_type([1, 2, 3, 4] | s, "pd.Series[bool]"), pd.Series, np.bool_)
-    check(assert_type(s | s2, "pd.Series[int]"), pd.Series, np.int_)
-    check(assert_type(s2 | s, "pd.Series[int]"), pd.Series, np.int_)
+    check(assert_type(s | s2, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(s2 | s, "pd.Series[int]"), pd.Series, np.integer)
 
     check(assert_type(s ^ [1, 2, 3, 4], "pd.Series[bool]"), pd.Series, np.bool_)
     check(assert_type([1, 2, 3, 4] ^ s, "pd.Series[bool]"), pd.Series, np.bool_)
-    check(assert_type(s ^ s2, "pd.Series[int]"), pd.Series, np.int_)
-    check(assert_type(s2 ^ s, "pd.Series[int]"), pd.Series, np.int_)
+    check(assert_type(s ^ s2, "pd.Series[int]"), pd.Series, np.integer)
+    check(assert_type(s2 ^ s, "pd.Series[int]"), pd.Series, np.integer)
 
 
 def test_logical_operators() -> None:
